refactor(central-server): separate config and implementations of adapters
Combining them into single modules cause a circular import, so
separating them is necessary.
