+++ CentralServer/centralserver/internals/adapters/config.py
@@ -1,12 +1,9 @@
-"""
-This module contains the database configuration classes.
-The actual database engine is created in the `db_handler.py` module.
-"""
-
 import os
 from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Any, override
+
+##### Database Adapter Configurations #####
 
 
 class DatabaseAdapterConfig(ABC):
@@ -137,3 +134,71 @@ class MySQLDatabaseConfig(DatabaseAdapterConfig):
     @override
     def connect_args(self) -> dict[str, Any]:
         return self._connect_args
+
+
+##### Object Store Adapter Configurations #####
+
+
+class ObjectStoreAdapterConfig(ABC):
+    """Adapter configuration for object store."""
+
+    @property
+    @abstractmethod
+    def info(self) -> dict[str, Any]:
+        """Get the object store adapter information"""
+
+
+class LocalObjectStoreAdapterConfig(ObjectStoreAdapterConfig):
+    """Adapter configuration for local object store."""
+
+    def __init__(self, filepath: str | None = None) -> None:
+        self.filepath: Path = Path(filepath or os.path.join(os.getcwd(), "data"))
+
+    @property
+    @override
+    def info(self) -> dict[str, Any]:
+        """Get the object store adapter information."""
+
+        return {
+            "name": "Local",
+            "filepath": str(self.filepath),
+        }
+
+
+class MinIOObjectStoreAdapterConfig(ObjectStoreAdapterConfig):
+    """Adapter configuration for MinIO."""
+
+    def __init__(
+        self,
+        access_key: str | None = None,
+        secret_key: str | None = None,
+        endpoint: str | None = None,
+        secure: bool | None = None,
+    ):
+        """Configuration for MinIO object store adapter.
+
+        Args:
+            access_key: The access key for MinIO. (required)
+            secret_key: The secret key for MinIO. (required)
+            endpoint: The URL of the MinIO server. (default: http://localhost:9000)
+            secure: Use secure (TLS) connection. (default: False).
+        """
+
+        if access_key is None or secret_key is None:
+            raise ValueError("The access key and secret key are required.")
+
+        self.access_key: str = access_key
+        self.secret_key: str = secret_key
+        self.endpoint: str = endpoint or "http://localhost:9000"
+        self.secure: bool = secure or False
+
+    @property
+    @override
+    def info(self) -> dict[str, Any]:
+        return {
+            "name": "MinIO",
+            "access_key_set": self.access_key != "",
+            "secret_key_set": self.secret_key != "",
+            "endpoint": self.endpoint,
+            "secure": self.secure,
+        }
